+++ main.py
@@ -16,10 +16,17 @@ import typing
 from functools import partial
 from pathlib import Path
 
-from src import GETTING_STARTED, exceptions, pre_flight_checks, run_migrate, run_reconcile, types_
-from src.clients import get_clients
-from src.constants import DEFAULT_BRANCH
-from src.types_ import ActionResult, PullRequestAction
+from gatekeeper import (
+    GETTING_STARTED,
+    exceptions,
+    pre_flight_checks,
+    run_migrate,
+    run_reconcile,
+    types_,
+)
+from gatekeeper.clients import get_clients
+from gatekeeper.constants import DEFAULT_BRANCH
+from gatekeeper.types_ import ActionResult, PullRequestAction
 
 GITHUB_HEAD_REF_ENV_NAME = "GITHUB_HEAD_REF"
 GITHUB_OUTPUT_ENV_NAME = "GITHUB_OUTPUT"
@@ -45,6 +52,7 @@ def _parse_env_vars() -> types_.UserInputs:
     github_access_token = os.getenv("INPUT_GITHUB_TOKEN")
     base_branch = os.getenv("INPUT_BASE_BRANCH", DEFAULT_BRANCH)
     commit_sha = os.getenv("INPUT_COMMIT_SHA")
+    charm_dir = os.getenv("INPUT_CHARM_DIR", "")
 
     event_path = os.getenv("GITHUB_EVENT_PATH")
     if not event_path:
@@ -79,6 +87,7 @@ def _parse_env_vars() -> types_.UserInputs:
         github_access_token=github_access_token,
         commit_sha=commit_sha,
         base_branch=base_branch,
+        charm_dir=charm_dir,
     )