+++ tests/factories.py
@@ -11,8 +11,8 @@ from typing import Generic, TypeVar
 
 import factory
 
-from src import index, types_
-from src.constants import DEFAULT_BRANCH
+from gatekeeper import index, types_
+from gatekeeper.constants import DEFAULT_BRANCH
 
 from . import types
 
@@ -359,6 +359,7 @@ class UserInputsFactory(
     base_branch = DEFAULT_BRANCH
     dry_run = False
     delete_pages = False
+    charm_dir = ""
 
 
 class TableRowFactory(