+++ tests/unit/test_migration/test_public.py
@@ -9,7 +9,7 @@ from unittest import mock
 
 import pytest
 
-from src import discourse, exceptions, migration, types_
+from gatekeeper import discourse, exceptions, migration, types_
 
 from ... import factories