+++ tests/unit/test_navigation_table.py
@@ -8,8 +8,8 @@
 
 import pytest
 
-from src import exceptions, navigation_table, types_
-from src.exceptions import NavigationTableParseError
+from gatekeeper import exceptions, navigation_table, types_
+from gatekeeper.exceptions import NavigationTableParseError
 
 from .. import factories
 from .helpers import assert_substrings_in_string